rerun on name prism instead of original lsTM
